Zack's review comments addressed
