+++ pilot/pkg/model/context.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/gogo/protobuf/types"
 	multierror "github.com/hashicorp/go-multierror"
+	
 	meshconfig "istio.io/api/mesh/v1alpha1"
 )
